Complete Timer for Alert
Add an interval parameter to the timer reducer through which the
interval between alerts is set.  This will complete the alert timer.

The user will be allowed to set the interval via a different function.
Also, the alert will only fire when the app is in focus.  Alerting from
the background will be implemented next.