Add Timer Interval to State
Add a slider to the home component that allows the user to select the
interval for the alert timer.  Add a variable to carry the interval to
the set-timer actions.  Add actions to remove the timers.  Add a default
interval to the add-user actions.

Also, add an array to the timer reducer that can carry the result of the
setTimeout function.  Add appropriate actions for appending that result
to that array and actions to remove them when they are redundant.