Add Welcome Screen
Add a Welcome Screen that allows the user to select whether she is a
patient or a standby-team member.  Add a variable to the state that will
store the selection and the accompanying actions.  Place the screen in
navigation so that it appears when the status has not yet been selected
(i.e., isPatient == null).