Add Firebase User to State
Add the Firebase User to state in via redux.  Use that user instead of 
fetching the user from Firebase auth.  Change the key for Firebase 
documents from a user's token to the user's Firebase uid because there 
needs to be a persistent identifier through which the standby will track 
check-ins.