Add Firestore Data to Standby Home
Add an action that grabs the data from a Firestore document (sign-in
time and check-in time) for a default patient.  Create a reducer to add
that data to state.  Add text to the home component for a standby user
that displays the sign-in and check-in time for the default patient.