Improve Timer Functions and Style
Make sure that the default timer interval is set only in the timer's
reducer.  Add a minimum and maximum value for the interval in the
timer's reducer.  Add a variable to the check-in function that logs the
interval to the Firestore.

Also, make sure all buttons have a style applied to them and that all
styles set a margin.