Work on Navigation
Navigation in the main component needs to be overhauled.  The app
container should switch between a loading screen, an authorization
stack, and application stacks for both patient and standby.  The welcome
component should be moved into an authorization stack along with the
patient and standby sign-in screens.  Navigation in the welcome
component and the action creators will need to be updated appropriately.
Remove the common home and sign-in components because they will no
longer be needed.