Adjust Actions to Fetch State from Firestore
Roll the getDocument function into setListener so that the state is
updated every time it is called.  Also, do not pass the database as a
function parameter; it is available from the import statement.  Finally,
adjust the passing of other variables so that this happens only when
necessary.  Make sure the setListener function passes its tests.