Return Null When Alert Fires
When the setListener and the setTimer functions fire an alert, there is
no need to store a new timer ID, which was not implemented correctly
anyway.  It should return null instead, and that should be dealt with in
the reducers.

Also, multiple alerts were firing because the timers were not being
removed because the fulfillment action had no variable.