Continue Work on checkCheckins Function
Integrate the setTimer function from the master branch into the
checkCheckins function, which was renamed from scheduledFunction.

Also, grab the Firestore collection instead of a single document, which
is no longer necessary since triggers are no longer being used.  Delete
the user document and just use the users collection, which eliminates
the need to change the action creators in the master branch with the
exception of adding the registration token when the documents are
created.