Improve Cloud Function checkCheckins
First, since getting registration tokens returns in a promise, it can
return an array with a single value of null.  Filter the resulting array
for that case in the then-statement.

Second, the case when the user has entered no alert times has not been
handled.  Handle it.

Lastly, remove comments that are no longer needed.